为近战武器增加移动动画支持 & 将put_away计入切换冷却 & 新增棒球棍和匕首

+++ src/main/java/me/xjqsh/lrtactical/item/melee/CombatData.java
@@ -33,7 +33,7 @@ public class CombatData {
             float factor,
 
             @SerializedName("knockback")
-            double knockback,
+            float knockback,
 
             @SerializedName("cooldown")
             int cooldown,
